Pulled newItems from Api, rendered them inside a carousel, and added a loading state, added Countdown.jsx componenet to handle expiryData.
